+++ modules/fortigate/fgt_ha/output.tf
@@ -1,0 +1,3 @@
+output "instance_id_list" {
+  value = google_compute_instance_group.fgt_umigs.*.id
+}