+++ modules/fortinet/generic_vm_standalone/versions.tf
@@ -1,0 +1,8 @@
+terraform {
+  required_version = ">=0.13, < 2.0.0"
+  required_providers {
+    google = {
+      version = ">= 5.0, <7.0"
+    }
+  }
+}